requirements.txt
torch==2.1.1
transformers==4.39.3
diffusers==0.21.1
torchvision==0.16.1
torchaudio==2.1.1
accelerate>=0.18.0
scikit-image
tqdm
numpy
pillow
matplotlib
scipy
einops
xformers
safetensors
pyyaml
ipython
omegaconf
opencv-python
tensorboard